--- app/model.js
@@ -1,4 +1,0 @@
-export const model = {
-    username: "donor1",
-    amountbloodL: 5,
-}
