+++ src/presenters/requestPresenter.jsx
@@ -3,8 +3,5 @@ import { observer } from "mobx-react-lite"
 import { RequestView } from "src/views/requestView"
 
 export const Request = observer((props) => {
-  return (
-    <RequestView
-    />
-  )
+  return <RequestView />
 })
