+++ src/views/requestView.jsx
@@ -1,16 +1,16 @@
-import {Text, StyleSheet, View} from "react-native"
+import { StyleSheet, Text, View } from "react-native"
 
-export function RequestView () {
-    return (
-        <View style = {styles.container}>
-        <Text>ththnbjh</Text>
-        </View>
-    )
+export function RequestView() {
+  return (
+    <View style={styles.container}>
+      <Text>ththnbjh</Text>
+    </View>
+  )
 }
 
-const styles = StyleSheet.create ({
-    container: {
-        backgroundColor: "#780000",
-        flex: 1,
-    },
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "#780000",
+    flex: 1,
+  },
 })
